# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PM Gestion API
  version: 1.1.0
  description: API de gestion de projets — Authentification JWT, RBAC 2-tier, real-time Socket.IO
  contact:
    name: TripleV
    url: https://github.com/TripleV-coder/Project-Manager
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: '{baseUrl}/api'
    description: Production
    variables:
      baseUrl:
        default: https://your-domain.com
security:
  - BearerAuth: []
  - CookieAuth: []
...
paths:
...
  /projects/{id}:
    get:
      tags: [Projects]
      summary: Get project details
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Project' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Projects]
      summary: Update a project
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectInput' }
      responses:
        '200': { description: Project updated }
    delete:
      tags: [Projects]
      summary: Delete a project
      parameters:
        - { $ref: '#/components/parameters/ResourceId' }
      responses:
        '204': { description: Project deleted }

  # ── Tasks ─────────────────────────────────────────────
...
  /sprints/{id}:
    get:
      tags: [Sprints]
      summary: Get sprint details
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sprint' }
    put:
      tags: [Sprints]
      summary: Update a sprint
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '200': { description: Sprint updated }
    delete:
      tags: [Sprints]
      summary: Delete a sprint
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '204': { description: Sprint deleted }
  /sprints/{id}/start:
    post:
      tags: [Sprints]
      summary: Start a sprint
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses:
        '200': { description: Sprint started }
...
  /timesheets:
    get:
      tags: [Timesheets]
      summary: List timesheets
      parameters:
        - { name: projet_id, in: query, schema: { type: string } }
        - { name: date_start, in: query, schema: { type: string, format: date } }
        - { name: date_end, in: query, schema: { type: string, format: date } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedResponse' }
    post:
      tags: [Timesheets]
      summary: Create a timesheet entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projet_id, date, heures]
              properties:
                projet_id: { type: string }
                task_id: { type: string }
                sprint_id: { type: string }
                date: { type: string, format: date }
                heures: { type: number, minimum: 0.25, maximum: 24 }
                description: { type: string }
      responses:
        '201': { description: Timesheet entry created }
...
  /push/subscribe:
    post:
      tags: [Push]
      summary: Register a push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscription: { type: object }
      responses:
        '200': { description: Subscription registered }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-f0-9]{24}$'
      description: MongoDB ObjectId
  responses:
...
    NotFound:
      description: Resource not found
...
  schemas:
    User:
      type: object
      properties:
        _id: { type: string }
        nom_complet: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        statut: { type: string, enum: [Actif, Inactif, Suspendu] }
        département: { type: string }
        poste: { type: string }
        compétences: { type: array, items: { type: string } }
    Role:
      type: object
      properties:
        _id: { type: string }
        nom: { type: string }
        description: { type: string }
        permissions:
          type: object
          description: 32 atomic boolean permissions
    Project:
      type: object
      properties:
        _id: { type: string }
        nom: { type: string }
        description: { type: string }
        statut: { type: string, enum: [Planification, En cours, En pause, Terminé, Annulé] }
        priorité: { type: string, enum: [Basse, Moyenne, Haute, Critique] }
        template_id: { type: string }
        chef_projet: { $ref: '#/components/schemas/User' }
        membres:
          type: array
          items:
            type: object
            properties:
              user_id: { type: string }
              project_role_id: { type: string }
        budget:
          type: object
          properties:
            prévisionnel: { type: number }
            réel: { type: number }
            devise: { type: string, default: FCFA }
        date_début: { type: string, format: date }
        date_fin_prévue: { type: string, format: date }
    ProjectInput:
      type: object
      required: [nom, template_id, chef_projet]
      properties:
        nom: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string }
        template_id: { type: string }
        chef_projet: { type: string }
        statut: { type: string }
        priorité: { type: string }
        date_début: { type: string, format: date }
        date_fin_prévue: { type: string, format: date }
...
    Sprint:
      type: object
      properties:
        _id: { type: string }
        nom: { type: string }
        objectif: { type: string }
        projet_id: { type: string }
        statut: { type: string, enum: [Planifié, Actif, Terminé] }
        date_début: { type: string, format: date }
        date_fin: { type: string, format: date }
        capacité_équipe: { type: number }
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean }
        data: { type: array, items: { type: object } }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
        hasNextPage: { type: boolean }